refactor(manifest-conformance): one gst parser
`gstPluginBaseName` and `readGstPluginDir` sat inside the `media-capabilities`
RULE, and `gst-plugins.mjs` imported them from there — so both bundle builders
and the on-target probe enrolled a conformance rule into the registry as a side
effect of asking what a plugin file is called. `defineRule` runs at module
scope; nothing in a builder wants that.

They move to `lib/gst-payload.mjs`, beside `binary.mjs` and for its reason: a
parser three readers share is not one caller's private detail.

`isGstElementName` joins them because the second copy had already drifted. The
rule accepted `^[a-z0-9_]+$` and `gtk-runtime-bundle-gates` asserted
`^[a-z0-9]+$` over the same field, so `avdec_aac` — the element ADR 0053 names
as the one that would close the AAC gap — was legal in the manifest and illegal
in the test.
